Add levee demo, add (leaky)drain demo, draft CCD, fix bugs.
Add: draft of raycasting based collision detection(CCD).

Fix: FluidSystem::addPointReuse(); f->nextFluidIndex was set incorrectly.
This caused intermittent lockups with marching cubes.

Fix: major memory leak; areSpheresGenerated was not set to true.
(thanks to majestik666 for the report)


Regarding the drain demo leak:
At present, fluid particles are pushed out of btCollsionObjects
when a collision is detected. The direction in which they are
accelerated is the shortest path that will cause them to stop
colliding.

As a result, if 2 btCollisionObjects are adjacent, with
no gap between, a colliding particle may be pushed
back and forth until it tunnels through.